requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
requests>=2.31.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
